change main component mycontent/saved
