+++ client/src/pages/MyContent/ContentData.jsx
@@ -1,0 +1,180 @@
+import React, { useState } from "react";
+
+function ContentData() {
+  const [contentData, setContentData] = useState({
+    articles: [
+      {
+        id: 1,
+        title: "The Future of Web Development",
+        preview: "Exploring emerging trends in modern web development...",
+        date: "2023-06-15T14:30:00",
+        views: 5245,
+        likes: 342,
+        comments: 28,
+        category: "Technology",
+        status: "published",
+        readTime: "4 min",
+        image: "https://source.unsplash.com/random/600x400/?web,development",
+        isFeatured: true,
+        tags: ["web", "trends", "javascript"],
+      },
+      {
+        id: 2,
+        title: "Sustainable UI Design Practices",
+        preview: "How to create eco-friendly digital interfaces...",
+        date: "2023-06-10T09:15:00",
+        views: 3856,
+        likes: 1024,
+        comments: 42,
+        category: "Design",
+        status: "published",
+        readTime: "6 min",
+        image: "https://source.unsplash.com/random/600x400/?design,ui",
+        isFeatured: false,
+        tags: ["design", "sustainability"],
+      },
+      {
+        id: 3,
+        title: "Draft: React Performance Optimization",
+        preview: "Advanced techniques to speed up your React apps...",
+        date: "2023-06-05T11:20:00",
+        views: 0,
+        likes: 0,
+        comments: 0,
+        category: "Development",
+        status: "draft",
+        readTime: "8 min",
+        image: "https://source.unsplash.com/random/600x400/?react,code",
+        isFeatured: false,
+        tags: ["react", "performance"],
+      },
+    ],
+    videos: [
+      {
+        id: 4,
+        title: "Building a Full-Stack App with Next.js",
+        preview: "Step-by-step tutorial on creating a modern web app...",
+        date: "2023-06-12T16:45:00",
+        views: 8421,
+        likes: 1254,
+        comments: 87,
+        category: "Tutorial",
+        status: "published",
+        duration: "24:32",
+        thumbnail: "https://source.unsplash.com/random/600x400/?coding,nextjs",
+        watchedPercent: 65,
+        tags: ["nextjs", "fullstack"],
+      },
+      {
+        id: 5,
+        title: "CSS Grid Deep Dive",
+        preview: "Mastering layout techniques with CSS Grid...",
+        date: "2023-06-08T13:25:00",
+        views: 6203,
+        likes: 2103,
+        comments: 87,
+        category: "Tutorial",
+        status: "published",
+        duration: "32:10",
+        thumbnail: "https://source.unsplash.com/random/600x400/?css,grid",
+        watchedPercent: 45,
+        tags: ["css", "layout"],
+      },
+      {
+        id: 6,
+        title: "Unlisted: TypeScript Best Practices",
+        preview: "Professional patterns for TypeScript development...",
+        date: "2023-06-01T10:15:00",
+        views: 153,
+        likes: 42,
+        comments: 5,
+        category: "Tutorial",
+        status: "unlisted",
+        duration: "18:45",
+        thumbnail:
+          "https://source.unsplash.com/random/600x400/?typescript,code",
+        watchedPercent: 72,
+        tags: ["typescript", "best-practices"],
+      },
+    ],
+    live: [
+      {
+        id: 7,
+        title: "Q&A: Ask Me Anything About React",
+        preview: "Live session answering your React questions...",
+        date: "2023-06-18T14:00:00",
+        views: 12453,
+        likes: 2845,
+        comments: 423,
+        category: "Live",
+        status: "ended",
+        duration: "1:25:10",
+        thumbnail: "https://source.unsplash.com/random/600x400/?live,stream",
+        isUpcoming: false,
+        tags: ["react", "q&a"],
+      },
+      {
+        id: 8,
+        title: "Upcoming: Building a Design System",
+        preview: "Live coding session creating a component library...",
+        date: "2023-06-25T15:30:00",
+        views: 0,
+        likes: 0,
+        comments: 0,
+        category: "Live",
+        status: "scheduled",
+        duration: null,
+        thumbnail: "https://source.unsplash.com/random/600x400/?design,system",
+        isUpcoming: true,
+        tags: ["design", "components"],
+      },
+    ],
+    opinions: [
+      {
+        id: 9,
+        title: "Why Remote Work is Here to Stay",
+        preview: "My perspective on the future of work arrangements...",
+        date: "2023-06-14T11:20:00",
+        views: 3842,
+        likes: 842,
+        comments: 128,
+        category: "Business",
+        status: "published",
+        image: "https://source.unsplash.com/random/600x400/?office,remote",
+        isFeatured: true,
+        tags: ["remote-work", "future"],
+      },
+      {
+        id: 10,
+        title: "The Ethics of AI Content Generation",
+        preview: "Should AI-generated content be credited differently?...",
+        date: "2023-06-07T17:30:00",
+        views: 1532,
+        likes: 456,
+        comments: 84,
+        category: "Technology",
+        status: "published",
+        image: "https://source.unsplash.com/random/600x400/?ai,ethics",
+        isFeatured: false,
+        tags: ["ai", "ethics"],
+      },
+      {
+        id: 11,
+        title: "Draft: Minimalism in Digital Products",
+        preview: "The value of simplicity in interface design...",
+        date: "2023-06-03T09:45:00",
+        views: 0,
+        likes: 0,
+        comments: 0,
+        category: "Design",
+        status: "draft",
+        image: "https://source.unsplash.com/random/600x400/?minimal,design",
+        isFeatured: false,
+        tags: ["design", "minimalism"],
+      },
+    ],
+  });
+  return contentData;
+}
+
+export default ContentData;
