+++ client/src/pages/MyContent/SelectdItem.jsx
@@ -1,0 +1,71 @@
+import React, { memo } from "react";
+import {
+  BookOpen,
+  Video,
+  Play,
+  MessageCircle,
+  Plus,
+  Filter,
+  Search,
+  Download,
+  Share2,
+  Edit,
+  Trash2,
+  MoreVertical,
+  X,
+  Eye,
+  Heart,
+  MessageSquare,
+  Calendar,
+  Bookmark,
+  Zap,
+  ChevronDown,
+  BarChart2,
+  TrendingUp,
+  Clock,
+  Users,
+  Tag,
+  CheckCircle,
+  AlertCircle,
+  Star,
+} from "lucide-react";
+function SelectdItem(props) {
+  const { selectedItems, setSelectedItems, activeTab, deleteSelected } = props;
+  return (
+    <div className="bg-blue-50 dark:bg-blue-900/30 rounded-lg p-3 mb-4 flex justify-between items-center">
+      <div className="flex items-center gap-3">
+        <span className="text-sm font-medium text-blue-800 dark:text-blue-200">
+          {selectedItems.length} item{selectedItems.length !== 1 ? "s" : ""}{" "}
+          selected
+        </span>
+        <button
+          onClick={() => setSelectedItems([])}
+          className="text-blue-600 dark:text-blue-400 hover:text-blue-800 dark:hover:text-blue-300 text-sm"
+        >
+          Clear selection
+        </button>
+      </div>
+      <div className="flex gap-2">
+        <button className="flex items-center gap-1 text-gray-600 dark:text-gray-400 hover:text-gray-800 dark:hover:text-gray-300 text-sm px-3 py-1.5 bg-white dark:bg-gray-700 rounded-lg border border-gray-300 dark:border-gray-600">
+          <Share2 size={14} />
+          <span>Share</span>
+        </button>
+        {activeTab !== "live" && (
+          <button className="flex items-center gap-1 text-gray-600 dark:text-gray-400 hover:text-gray-800 dark:hover:text-gray-300 text-sm px-3 py-1.5 bg-white dark:bg-gray-700 rounded-lg border border-gray-300 dark:border-gray-600">
+            <Edit size={14} />
+            <span>Edit</span>
+          </button>
+        )}
+        <button
+          onClick={deleteSelected}
+          className="flex items-center gap-1 text-red-600 dark:text-red-400 hover:text-red-800 dark:hover:text-red-300 text-sm px-3 py-1.5 bg-white dark:bg-gray-700 rounded-lg border border-gray-300 dark:border-gray-600"
+        >
+          <Trash2 size={14} />
+          <span>Delete</span>
+        </button>
+      </div>
+    </div>
+  );
+}
+
+export default memo(SelectdItem);
